pyproject.toml
[project]
name = "polarized"
version = "0.1.0"
description = ""
authors = [
    {name = "javierlnmn",email = "jlm03123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "asgiref (==3.8.1)",
    "cachetools (==5.5.0)",
    "certifi (==2024.12.14)",
    "channels (==4.2.0)",
    "charset-normalizer (==3.4.1)",
    "django (==5.0.6)",
    "django-4-jet (==1.0.9)",
    "django-browser-reload (==1.13.0)",
    "django-modeltranslation (==0.19.12)",
    "django-tailwind (==3.8.0)",
    "google-api-core (==2.24.0)",
    "google-api-python-client (==2.157.0)",
    "google-auth (==2.37.0)",
    "google-auth-httplib2 (==0.2.0)",
    "googleapis-common-protos (==1.66.0)",
    "httplib2 (==0.22.0)",
    "idna (==3.10)",
    "mysqlclient (==2.2.6)",
    "proto-plus (==1.25.0)",
    "protobuf (==5.29.2)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.1)",
    "pyparsing (==3.2.1)",
    "python-dotenv (==1.0.1)",
    "python-slugify (==8.0.4)",
    "requests (==2.32.3)",
    "rsa (==4.9)",
    "six (==1.16.0)",
    "sqlparse (==0.5.1)",
    "text-unidecode (==1.3)",
    "tzdata (==2024.2)",
    "uritemplate (==4.1.1)",
    "urllib3 (==2.3.0)",
    "ruff (>=0.14.6,<0.15.0)",
    "pre-commit (>=4.5.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.ruff]
line-length = 110
exclude = ["migrations"]
lint.select = ["F", "E", "W", "I"]

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"
